+++ packages/stream_chat_persistence/lib/src/dao/channel_query_dao.g.dart
@@ -6,7 +6,7 @@ part of 'channel_query_dao.dart';
 // DaoGenerator
 // **************************************************************************
 
-mixin _$ChannelQueryDaoMixin on DatabaseAccessor<MoorChatDatabase> {
+mixin _$ChannelQueryDaoMixin on DatabaseAccessor<DriftChatDatabase> {
   $ChannelQueriesTable get channelQueries => attachedDatabase.channelQueries;
   $ChannelsTable get channels => attachedDatabase.channels;
   $UsersTable get users => attachedDatabase.users;