+++ packages/stream_chat_persistence/lib/src/dao/reaction_dao.g.dart
@@ -6,7 +6,7 @@ part of 'reaction_dao.dart';
 // DaoGenerator
 // **************************************************************************
 
-mixin _$ReactionDaoMixin on DatabaseAccessor<MoorChatDatabase> {
+mixin _$ReactionDaoMixin on DatabaseAccessor<DriftChatDatabase> {
   $ReactionsTable get reactions => attachedDatabase.reactions;
   $UsersTable get users => attachedDatabase.users;
 }