+++ packages/stream_chat_persistence/test/src/dao/member_dao_test.dart
@@ -3,14 +3,14 @@ import 'dart:math' as math;
 import 'package:flutter_test/flutter_test.dart';
 import 'package:stream_chat/stream_chat.dart';
 import 'package:stream_chat_persistence/src/dao/dao.dart';
-import 'package:stream_chat_persistence/src/db/moor_chat_database.dart';
+import 'package:stream_chat_persistence/src/db/drift_chat_database.dart';
 
 import '../../stream_chat_persistence_client_test.dart';
 import '../utils/date_matcher.dart';
 
 void main() {
   late MemberDao memberDao;
-  late MoorChatDatabase database;
+  late DriftChatDatabase database;
 
   setUp(() {
     database = testDatabaseProvider('testUserId');