+++ packages/stream_chat_persistence/test/src/dao/read_dao_test.dart
@@ -1,14 +1,14 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:stream_chat/stream_chat.dart';
 import 'package:stream_chat_persistence/src/dao/dao.dart';
-import 'package:stream_chat_persistence/src/db/moor_chat_database.dart';
+import 'package:stream_chat_persistence/src/db/drift_chat_database.dart';
 
 import '../../stream_chat_persistence_client_test.dart';
 import '../utils/date_matcher.dart';
 
 void main() {
   late ReadDao readDao;
-  late MoorChatDatabase database;
+  late DriftChatDatabase database;
 
   setUp(() {
     database = testDatabaseProvider('testUserId');