+++ packages/stream_chat_persistence/test/src/mapper/message_mapper_test.dart
@@ -3,7 +3,7 @@ import 'dart:math' as math;
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:stream_chat/stream_chat.dart';
-import 'package:stream_chat_persistence/src/db/moor_chat_database.dart';
+import 'package:stream_chat_persistence/src/db/drift_chat_database.dart';
 import 'package:stream_chat_persistence/src/mapper/message_mapper.dart';
 
 import '../utils/date_matcher.dart';