+++ packages/stream_chat_persistence/test/src/mapper/pinned_message_reaction_mapper_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:stream_chat/stream_chat.dart';
-import 'package:stream_chat_persistence/src/db/moor_chat_database.dart';
+import 'package:stream_chat_persistence/src/db/drift_chat_database.dart';
 import 'package:stream_chat_persistence/src/mapper/pinned_message_reaction_mapper.dart';
 
 import '../utils/date_matcher.dart';