+++ packages/stream_chat/CHANGELOG.md
@@ -1,8 +1,8 @@
-## Upcoming
+## 3.2.0
 
 🐞 Fixed
 
-- markAllRead() now updates local channel states.
+- `markAllRead()` now updates local channel states.
 - [[#744]](https://github.com/GetStream/stream-chat-flutter/issues/744) Fixed unread count not updating correctly
 
 ## 3.1.1